+++ hutool-core/src/test/java/cn/hutool/core/text/split/StrSpliterTest.java
@@ -60,6 +60,15 @@ public class StrSpliterTest {
 		String str = "";
 		final String[] split = str.split(",");
 		final String[] strings = StrSplitter.splitToArray(str, ",", -1, false, false);
+		Assertions.assertNotNull(strings);
 		Assertions.assertArrayEquals(split, strings);
+	}
+
+	@Test
+	public void splitNullTest(){
+		String str = null;
+		final String[] strings = StrSplitter.splitToArray(str, ",", -1, false, false);
+		Assertions.assertNotNull(strings);
+		Assertions.assertEquals(0, strings.length);
 	}
 }