+++ hutool-core/src/test/java/cn/hutool/core/util/StrUtilTest.java
@@ -425,12 +425,6 @@ public class StrUtilTest {
 				.set("HelloWorld_test", "hello_world_test")
 				.set("H2", "H2")
 				.set("H#case", "H#case")
-				.forEach((key, value) -> Assertions.assertEquals(value, StrUtil.toUnderlineCase(key)));
-	}
-
-	@Test
-	public void toUnderLineCaseTest2() {
-		Dict.create()
 				.set("PNLabel", "PN_label")
 				.forEach((key, value) -> Assertions.assertEquals(value, StrUtil.toUnderlineCase(key)));
 	}
@@ -533,10 +527,18 @@ public class StrUtilTest {
 
 	@Test
 	public void briefTest() {
-		String str = RandomUtil.randomString(1000);
-		int maxLength = RandomUtil.randomInt(1000);
-		String brief = StrUtil.brief(str, maxLength);
-		Assertions.assertEquals(brief.length(), maxLength);
+		// case: 1 至 str.length - 1
+		String str = RandomUtil.randomString(RandomUtil.randomInt(1, 100));
+		for (int maxLength = 1; maxLength < str.length(); maxLength++) {
+			String brief = StrUtil.brief(str, maxLength);
+			Assertions.assertEquals(brief.length(), maxLength);
+		}
+
+		// case: 不会格式化的值
+		Assertions.assertEquals(str, StrUtil.brief(str, 0));
+		Assertions.assertEquals(str, StrUtil.brief(str, -1));
+		Assertions.assertEquals(str, StrUtil.brief(str, str.length()));
+		Assertions.assertEquals(str, StrUtil.brief(str, str.length() + 1));
 	}
 
 	@Test
@@ -558,8 +560,21 @@ public class StrUtilTest {
 	@Test
 	public void briefTest3() {
 		String str = "123abc";
-		int maxLength = 3;
+
+		int maxLength = 6;
 		String brief = StrUtil.brief(str, maxLength);
+		Assertions.assertEquals(str, brief);
+
+		maxLength = 5;
+		brief = StrUtil.brief(str, maxLength);
+		Assertions.assertEquals("1...c", brief);
+
+		maxLength = 4;
+		brief = StrUtil.brief(str, maxLength);
+		Assertions.assertEquals("1..c", brief);
+
+		maxLength = 3;
+		brief = StrUtil.brief(str, maxLength);
 		Assertions.assertEquals("1.c", brief);
 
 		maxLength = 2;