+++ hutool-json/src/test/java/cn/hutool/json/Issue644Test.java
@@ -1,5 +1,6 @@
 package cn.hutool.json;
 
+import cn.hutool.core.date.LocalDateTimeUtil;
 import lombok.Data;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -19,10 +20,12 @@ public class Issue644Test {
 		final JSONObject jsonObject = JSONUtil.parseObj(beanWithDate);
 
 		BeanWithDate beanWithDate2 = JSONUtil.toBean(jsonObject, BeanWithDate.class);
-		Assertions.assertEquals(beanWithDate.getDate(), beanWithDate2.getDate());
+		Assertions.assertEquals(LocalDateTimeUtil.formatNormal(beanWithDate.getDate()),
+				LocalDateTimeUtil.formatNormal(beanWithDate2.getDate()));
 
 		beanWithDate2 = JSONUtil.toBean(jsonObject.toString(), BeanWithDate.class);
-		Assertions.assertEquals(beanWithDate.getDate(), beanWithDate2.getDate());
+		Assertions.assertEquals(LocalDateTimeUtil.formatNormal(beanWithDate.getDate()),
+				LocalDateTimeUtil.formatNormal(beanWithDate2.getDate()));
 	}
 
 	@Data